requirements.txt
# Core
Django==3.2.9
gunicorn==20.1.0

# For the persistence stores
psycopg2-binary==2.9.2
dj-database-url==0.5.0

# For Pub/Sub
redis==4.0.2

# Rest apis
djangorestframework==3.12.4
Markdown==3.3.6
django-filter==21.1

# Developer Tools
ipdb==0.13.9
ipython==7.29.0
django-extensions==3.1.3
mkdocs==1.2.3
flake8==4.0.1
black==21.11b1

# Testing
mock==4.0.3
factory-boy==3.2.1
coverage==6.1.2

# Environment variables
django-environ==0.8.1
